Tests: split off two tests for the Runner class
... to their own test classes using the new `AbstractRunnerTestCase` base class.
