+++ tests/Core/Ruleset/ExplainTest.php
@@ -10,7 +10,6 @@
 namespace PHP_CodeSniffer\Tests\Core\Ruleset;
 
 use PHP_CodeSniffer\Ruleset;
-use PHP_CodeSniffer\Runner;
 use PHP_CodeSniffer\Tests\ConfigDouble;
 use PHPUnit\Framework\TestCase;
 
@@ -208,57 +207,6 @@ final class ExplainTest extends TestCase
         $ruleset->explain();
 
     }//end testExplainWithDeprecatedSniffs()
-
-
-    /**
-     * Test that each standard passed on the command-line is explained separately.
-     *
-     * @covers \PHP_CodeSniffer\Runner::runPHPCS
-     *
-     * @return void
-     */
-    public function testExplainWillExplainEachStandardSeparately()
-    {
-        if (PHP_CODESNIFFER_CBF === true) {
-            $this->markTestSkipped('This test needs CS mode to run');
-        }
-
-        $standard        = __DIR__.'/ExplainSingleSniffTest.xml';
-        $_SERVER['argv'] = [
-            'phpcs',
-            '-e',
-            "--standard=PSR1,$standard",
-            '--report-width=80',
-        ];
-
-        $expected  = PHP_EOL;
-        $expected .= 'The PSR1 standard contains 8 sniffs'.PHP_EOL.PHP_EOL;
-        $expected .= 'Generic (4 sniffs)'.PHP_EOL;
-        $expected .= '------------------'.PHP_EOL;
-        $expected .= '  Generic.Files.ByteOrderMark'.PHP_EOL;
-        $expected .= '  Generic.NamingConventions.UpperCaseConstantName'.PHP_EOL;
-        $expected .= '  Generic.PHP.DisallowAlternativePHPTags'.PHP_EOL;
-        $expected .= '  Generic.PHP.DisallowShortOpenTag'.PHP_EOL.PHP_EOL;
-        $expected .= 'PSR1 (3 sniffs)'.PHP_EOL;
-        $expected .= '---------------'.PHP_EOL;
-        $expected .= '  PSR1.Classes.ClassDeclaration'.PHP_EOL;
-        $expected .= '  PSR1.Files.SideEffects'.PHP_EOL;
-        $expected .= '  PSR1.Methods.CamelCapsMethodName'.PHP_EOL.PHP_EOL;
-        $expected .= 'Squiz (1 sniff)'.PHP_EOL;
-        $expected .= '---------------'.PHP_EOL;
-        $expected .= '  Squiz.Classes.ValidClassName'.PHP_EOL.PHP_EOL;
-
-        $expected .= 'The ExplainSingleSniffTest standard contains 1 sniff'.PHP_EOL.PHP_EOL;
-        $expected .= 'Squiz (1 sniff)'.PHP_EOL;
-        $expected .= '---------------'.PHP_EOL;
-        $expected .= '  Squiz.Scope.MethodScope'.PHP_EOL;
-
-        $this->expectOutputString($expected);
-
-        $runner = new Runner();
-        $runner->runPHPCS();
-
-    }//end testExplainWillExplainEachStandardSeparately()
 
 
 }//end class
